finished all query and insert functions and exported to test.js
